LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/text/2d.text.align.invalid-expected.txt

2d.text.align.invalid


FAIL OffscreenCanvas test: 2d.text.align.invalid assert_equals: ctx.textAlign === 'start' (got bogus[string], expected start[string]) expected "start" but got "bogus"
